requirements.txt
fastapi[all]
openai
python-dotenv
google-genai
Pillow
fastmcp
tavily-python
ngrok
pydantic-ai
